refactor: rename Navbar
