+++ client/components/UserHome.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import MobileFooter from './MobileFooter'
+import MobileFooter from './mobile-footer'
 
 /**
  * COMPONENT
