+++ client/components/mobile-footer.js
